fix observations and integrations with header feature
