+++ src/main/java/org/umssdiplo/automationv01/core/managepage/SHLogin.java
@@ -3,38 +3,41 @@ package org.umssdiplo.automationv01.core.managepage;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.umssdiplo.automationv01.core.utils.CommonEvents;
+import org.umssdiplo.automationv01.core.utils.PropertyAccessor;
 
 public class SHLogin extends BasePage {
+    @FindBy(name = "username")
+    private WebElement usernameInputField;
 
-    @FindBy(name = "org_button")
-    private WebElement organizationTab;
+    @FindBy(name = "password")
+    private WebElement passwordInputField;
 
-    @FindBy(name = "emp_button")
-    private WebElement employeeTab;
+    @FindBy(id = "login")
+    private WebElement aceptLoginBtn;
 
-    @FindBy(name = "accidents_button")
-    private WebElement accidentsTab;
+    @FindBy(name = "close_login_button")
+    private WebElement closeLoginBtn;
 
-    @FindBy(name = "aassign_button")
-    private WebElement assignTab;
-
-    public SHOrganization clickOrganizationTab(){
-        CommonEvents.clickButton(organizationTab);
-        return new SHOrganization();
+    public void setUsernameInputField() {
+        CommonEvents.setInputField(usernameInputField, PropertyAccessor.getInstance().getUser());
     }
 
-    public SHEmployee clickEmployeeTab(){
-        CommonEvents.clickButton(employeeTab);
-        return new SHEmployee();
+    public void setPasswordInputField() {
+        CommonEvents.setInputField(passwordInputField, PropertyAccessor.getInstance().getPassword());
     }
 
-    public SHAccident clickAccidentTab(){
-        CommonEvents.clickButton(employeeTab);
-        return new SHAccident();
+    public void clickAcceptLoginButton() {
+        CommonEvents.clickButton(aceptLoginBtn);
     }
 
-    public SHAssignation clickAssignTab(){
-        CommonEvents.clickButton(employeeTab);
-        return new SHAssignation();
+    public void clickCloseLoginButton() {
+        CommonEvents.clickButton(closeLoginBtn);
+    }
+
+    public void fillCredentials() {
+        setUsernameInputField();
+        setPasswordInputField();
+        clickAcceptLoginButton();
+        clickCloseLoginButton();
     }
 }
