+++ src/test/java/org/umssdiplo/automationv01/stepdefinitionproject/LoginSteps.java
@@ -2,18 +2,21 @@ package org.umssdiplo.automationv01.stepdefinitionproject;
 
 import cucumber.api.java.en.And;
 import cucumber.api.java.en.Given;
-import org.umssdiplo.automationv01.core.managepage.SmarthHouse.SHLogin;
+import org.umssdiplo.automationv01.core.managepage.HeaderWithLogin;
+import org.umssdiplo.automationv01.core.managepage.SHLogin;
 import org.umssdiplo.automationv01.core.utils.LoadPage;
 
 public class LoginSteps {
     private SHLogin login;
+    private HeaderWithLogin headerWithLogin;
 
-    @Given("^Given I loging to 'SMARTHOUSE' page with properly credentials$")
+    @Given("^Given I loging to 'SMARTHOUSE' page")
     public void smarthouse_s_page_is_loaded() throws Throwable {
-        login = LoadPage.loginPage();
+        headerWithLogin = LoadPage.headerPageWithLogin();
+        login = headerWithLogin.openLoginPage();
     }
 
-    @And("^I fill credentials with admin user$")
+    @And("^I fill properly credentials with admin user$")
     public void fill_credentials_with_admin_user() throws Throwable {
         login.fillCredentials();
     }
