+++ src/test/resources/smarthHouseFeatures/Login.feature
@@ -2,5 +2,5 @@
 Feature: SHLogin
 
   Scenario: Home page is displayed once set credential in login page
-    Given Given I loging to 'SMARTHOUSE' page with properly credentials
-    And I fill credentials with admin user
+    Given Given I loging to 'SMARTHOUSE' page
+    And I fill properly credentials with admin user
